+++ breadlang/include/runtime/array_utils.h
@@ -3,8 +3,8 @@
 
 #include "runtime/runtime.h"
 
-struct BreadArray* bread_range_create(int start, int end, int step);
-struct BreadArray* bread_range(int n);
+struct BreadArray* bread_range_create(int64_t start, int64_t end, int64_t step);
+struct BreadArray* bread_range(int64_t n);
 int bread_array_get_value(struct BreadArray* a, int idx, BreadValue* out);
 int bread_value_array_get(BreadValue* array_val, int idx, BreadValue* out);
 int bread_value_array_length(BreadValue* array_val);
